admin forms part 2

+++ server/app/http/controllers/Admin/blog.controller.js
@@ -15,6 +15,9 @@ module.exports = new class AdminBlogController extends DefaultController {
             req.body.image = (path.join(req.body.fileUploadPath, req.body.fileName)).replaceAll(/\\/gi)
             const image = req.body.image
             const author = req?.user?._id
+            if (await BlogModel.findOne({title})) {
+                throw createHttpError.BadRequest("Blog with this title already exists")
+            }
             await BlogModel.create({title, overView, content, tags, category, image, author}).then(() => {
                 return res.status(StatusCodes.OK).json({
                     success: true,
@@ -73,9 +76,15 @@ module.exports = new class AdminBlogController extends DefaultController {
             if (!isValidObjectId(id)) {
                 throw createHttpError.BadRequest("the id is not valid")
             }
+            const blog = await BlogModel.findOne({_id: id}).catch(err => {
+                throw createHttpError.InternalServerError(err)
+            })
+            if (!blog) {
+                throw createHttpError.BadRequest("the blog could not be found")
+            }
             await BlogModel.deleteOne({_id: id}).then((result) => {
                 if (result.deletedCount > 0) {
-                    deleteImageFromPath(result.image)
+                    deleteImageFromPath(blog.image)
                     return res.status(StatusCodes.OK).json({
                         success: true,
                         message: "blog has been deleted successfully."
@@ -92,42 +101,15 @@ module.exports = new class AdminBlogController extends DefaultController {
 
     async getAllBlogs(req, res, next) {
         try {
-            const blogs = await BlogModel.aggregate([
-                {$match: {}},
-                {
-                    $lookup: {
-                        from: "users",
-                        foreignField: "_id",
-                        localField: "author",
-                        as: "author"
-                    }
-                },
-                {
-                    $unwind: "$author"
-                },
-                {
-                    $lookup: {
-                        from: "categories",
-                        foreignField: "_id",
-                        localField: "category",
-                        as: "category"
-                    }
-                },
-                {
-                    $unwind: "$category"
-                },
-                {
-                    $project: {
-                        "author.__v": 0,
-                        "category.__v": 0,
-                        "author.otp": 0,
-                        "author.Roles": 0,
-                        "author.discount": 0,
-                        "author.orders": 0,
-                    }
-                }
-            ]).catch(err => {
-                throw createHttpError.INTERNALSERVERERROR(err)
+            const blogs = await BlogModel.find({}, {
+
+                like: 0,
+                dislike: 0,
+                bookmarks: 0,
+                comments: 0,
+
+            }).catch(err => {
+                throw createHttpError.InternalServerError(err)
             })
             return res.status(StatusCodes.OK).json({
                 success: true,
